небольшой рефакторинг + создание компонента листа навигации
