+++ project/src/components/error-message/error-message.tsx
@@ -2,7 +2,7 @@ import {useAppSelector} from '../../hooks/useAppSelector';
 import './error-style.css';
 
 export default function ErrorMessage(): JSX.Element | null {
-  const error = useAppSelector((state) => state.error);
+  const {error} = useAppSelector((state) => state);
 
   return error ? <div className='error-message'>{error}</div> : null;
 }
